[refactor]: dto 에서 request 와 response 의 구분 없앰
